requirements-dev.txt
flake8
coverage
mypy
twine
boto3
google-cloud-storage >= 1.37.1
-r requirements.txt
